refactor: flatten dateUtils module
