+++ src/commands/reminder/remind-on-date.ts
@@ -1,6 +1,6 @@
 import { SlashCommandSubcommandBuilder } from 'discord.js';
 import { Result } from 'oxide.ts';
-import { convertDateToEpoch } from '../../utils/dateUtils';
+import { convertDateToEpoch } from '../../utils/date-utils';
 import type { CommandHandler, Subcommand } from '../builder';
 import { saveReminder } from './reminder-utils';
