+++ src/commands/reminder/update.ts
@@ -1,6 +1,6 @@
 import { SlashCommandSubcommandBuilder } from 'discord.js';
 import { Result } from 'oxide.ts';
-import { convertDateToEpoch } from '../../utils/dateUtils';
+import { convertDateToEpoch } from '../../utils/date-utils';
 import type { CommandHandler, Subcommand } from '../builder';
 import { updateReminder } from './reminder-utils';
