+++ src/scripts/delete-guild-commands.ts
@@ -1,6 +1,6 @@
 import { Result } from 'oxide.ts';
 import { deployGuildCommands } from '../commands/deploy-command';
-import { getCurrentUnixTime } from '../utils/dateUtils';
+import { getCurrentUnixTime } from '../utils/date-utils';
 import { loadEnv } from '../utils/loadEnv';
 import { logger } from '../utils/logger';
